Replace table_end() with table_find()
Instead of always inserting at the end of the table, use binary search
to find the correct index for the given key. Id the location returned
is not empty, throw a duplicate key error.
Update .btree test to expect keys to be ordered.
Add a test to check duplicate key error message.